LayoutTests/platform/efl/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
PASS rects[1].top is 19
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 59
PASS rects[2].width is 398
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 99
PASS rects[3].width is 360
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 139
PASS rects[4].width is 306
FAIL rects[4].height should be 18. Was 17.
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
PASS rects[0].top is 215
PASS rects[0].width is 396
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 255
PASS rects[1].width is 398
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 295
PASS rects[2].width is 360
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 335
PASS rects[3].width is 306
FAIL rects[3].height should be 18. Was 17.
Test 3
PASS rects.length is 4
PASS rects[0].left is 80
PASS rects[0].top is 411
PASS rects[0].width is 324
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 451
PASS rects[1].width is 398
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 491
PASS rects[2].width is 360
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 531
PASS rects[3].width is 88
FAIL rects[3].height should be 18. Was 17.
Test 4
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 596
PASS rects[0].width is 400
PASS rects[0].height is 278
PASS rects[1].left is 8
PASS rects[1].top is 607
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 725
PASS rects[2].width is 242
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 326
PASS rects[3].top is 725
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 765
PASS rects[4].width is 341
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
PASS rects[5].top is 805
PASS rects[5].width is 366
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
PASS rects[6].top is 845
PASS rects[6].width is 67
FAIL rects[6].height should be 18. Was 17.
Test 5
PASS rects.length is 5
PASS rects[0].left is 80
PASS rects[0].top is 921
PASS rects[0].width is 324
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 1039
PASS rects[1].width is 242
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 250
PASS rects[2].top is 950
PASS rects[2].width is 76
PASS rects[2].height is 103
PASS rects[3].left is 326
PASS rects[3].top is 1039
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 1079
PASS rects[4].width is 284
FAIL rects[4].height should be 18. Was 17.
Test 6
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 1224
PASS rects[0].width is 400
PASS rects[0].height is 200
PASS rects[1].left is 8
PASS rects[1].top is 1235
PASS rects[1].width is 396
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 1275
PASS rects[2].width is 57
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 105
PASS rects[3].top is 1275
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 1315
PASS rects[4].width is 387
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
PASS rects[5].top is 1355
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
PASS rects[6].top is 1395
PASS rects[6].width is 171
FAIL rects[6].height should be 18. Was 17.
Test 7
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 1471
PASS rects[0].width is 396
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 1500
PASS rects[1].width is 400
PASS rects[1].height is 40
PASS rects[2].left is 8
PASS rects[2].top is 1511
PASS rects[2].width is 57
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 105
PASS rects[3].top is 1511
PASS rects[3].width is 44
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 1551
PASS rects[4].width is 387
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
PASS rects[5].top is 1591
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
PASS rects[6].top is 1631
PASS rects[6].width is 171
FAIL rects[6].height should be 18. Was 17.
Test 8
PASS rects.length is 4
FAIL rects[0].left should be 168. Was 169.
PASS rects[0].top is 1678
FAIL rects[0].width should be 269. Was 268.
FAIL rects[0].height should be 316. Was 315.
FAIL rects[1].left should be 137. Was 138.
PASS rects[1].top is 1704
FAIL rects[1].width should be 271. Was 270.
FAIL rects[1].height should be 317. Was 316.
PASS rects[2].left is 107
PASS rects[2].top is 1729
PASS rects[2].width is 246
FAIL rects[2].height should be 289. Was 288.
FAIL rects[3].left should be 76. Was 77.
PASS rects[3].top is 1755
FAIL rects[3].width should be 212. Was 211.
FAIL rects[3].height should be 247. Was 246.
Test 9
FAIL rects.length should be 1. Was 0.
FAIL rects[0].left should be 8. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].left')
FAIL rects[0].top should be 1903. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].top')
FAIL rects[0].width should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].width')
FAIL rects[0].height should be 18. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].height')
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
PASS rects[0].top is 1903
PASS rects[0].width is 0
FAIL rects[0].height should be 18. Was 17.
Test 10
PASS rects.length is 4
PASS rects[0].left is 19
PASS rects[0].top is 2088
FAIL rects[0].width should be 18. Was 17.
PASS rects[0].height is 393
PASS rects[1].left is 59
PASS rects[1].top is 2088
FAIL rects[1].width should be 18. Was 17.
PASS rects[1].height is 377
PASS rects[2].left is 99
PASS rects[2].top is 2088
FAIL rects[2].width should be 18. Was 17.
PASS rects[2].height is 372
PASS rects[3].left is 139
PASS rects[3].top is 2088
FAIL rects[3].width should be 18. Was 17.
PASS rects[3].height is 360
Test 11
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
